Split ActionMenu stories into fixtures and examples
