+++ src/stories/ActionMenu2/examples.stories.tsx
@@ -1,0 +1,63 @@
+import React from 'react'
+import {Meta} from '@storybook/react'
+import {ThemeProvider} from '../..'
+import BaseStyles from '../../BaseStyles'
+import {ActionMenu} from '../../ActionMenu2'
+import {ActionList} from '../../ActionList2'
+
+const meta: Meta = {
+  title: 'Composite components/ActionMenu2/examples',
+  component: ActionMenu,
+  decorators: [
+    (Story: React.ComponentType): JSX.Element => (
+      <ThemeProvider>
+        <BaseStyles>
+          <Story />
+        </BaseStyles>
+      </ThemeProvider>
+    )
+  ],
+  parameters: {
+    controls: {
+      disabled: true
+    }
+  }
+}
+export default meta
+
+export function SimpleListStory(): JSX.Element {
+  const [actionFired, fireAction] = React.useState('')
+  const onSelect = (name: string) => fireAction(name)
+
+  return (
+    <>
+      <h1>Simple Menu</h1>
+      <h2>Last option activated: {actionFired}</h2>
+      <ActionMenu>
+        <ActionMenu.Button>Menu</ActionMenu.Button>
+        <ActionMenu.Overlay>
+          <ActionList>
+            <ActionList.Item onSelect={() => onSelect('Copy link')}>
+              Copy link
+              <ActionList.TrailingVisual>⌘C</ActionList.TrailingVisual>
+            </ActionList.Item>
+            <ActionList.Item onSelect={() => onSelect('Quote reply')}>
+              Quote reply
+              <ActionList.TrailingVisual>⌘Q</ActionList.TrailingVisual>
+            </ActionList.Item>
+            <ActionList.Item onSelect={() => onSelect('Edit comment')}>
+              Edit comment
+              <ActionList.TrailingVisual>⌘E</ActionList.TrailingVisual>
+            </ActionList.Item>
+            <ActionList.Divider />
+            <ActionList.Item variant="danger" onSelect={() => onSelect('Delete file')}>
+              Delete file
+              <ActionList.TrailingVisual>⌘D</ActionList.TrailingVisual>
+            </ActionList.Item>
+          </ActionList>
+        </ActionMenu.Overlay>
+      </ActionMenu>
+    </>
+  )
+}
+SimpleListStory.storyName = 'Simple Menu'
