+++ src/stories/ActionMenu2/fixtures.stories.tsx
@@ -1,16 +1,16 @@
 import React from 'react'
 import {Meta} from '@storybook/react'
-import {ThemeProvider} from '..'
-import BaseStyles from '../BaseStyles'
-import {ActionMenu} from '../ActionMenu2'
-import {ActionList} from '../ActionList2'
-import {Button} from '../Button2'
-import {IconButton} from '../Button2/IconButton'
-import Box from '../Box'
-import Text from '../Text'
-import TextInput from '../TextInput'
-import StyledOcticon from '../StyledOcticon'
-import FormGroup from '../FormGroup'
+import {ThemeProvider} from '../..'
+import BaseStyles from '../../BaseStyles'
+import {ActionMenu} from '../../ActionMenu2'
+import {ActionList} from '../../ActionList2'
+import {Button} from '../../Button2'
+import {IconButton} from '../../Button2/IconButton'
+import Box from '../../Box'
+import Text from '../../Text'
+import TextInput from '../../TextInput'
+import StyledOcticon from '../../StyledOcticon'
+import FormGroup from '../../FormGroup'
 import {
   ServerIcon,
   PlusCircleIcon,
@@ -29,7 +29,7 @@ import {
 } from '@primer/octicons-react'
 
 const meta: Meta = {
-  title: 'Composite components/ActionMenu2',
+  title: 'Composite components/ActionMenu2/fixtures',
   component: ActionMenu,
   decorators: [
     (Story: React.ComponentType): JSX.Element => (
@@ -47,43 +47,6 @@ const meta: Meta = {
   }
 }
 export default meta
-
-export function SimpleListStory(): JSX.Element {
-  const [actionFired, fireAction] = React.useState('')
-  const onSelect = (name: string) => fireAction(name)
-
-  return (
-    <>
-      <h1>Simple Menu</h1>
-      <h2>Last option activated: {actionFired}</h2>
-      <ActionMenu>
-        <ActionMenu.Button>Menu</ActionMenu.Button>
-        <ActionMenu.Overlay>
-          <ActionList>
-            <ActionList.Item onSelect={() => onSelect('Copy link')}>
-              Copy link
-              <ActionList.TrailingVisual>⌘C</ActionList.TrailingVisual>
-            </ActionList.Item>
-            <ActionList.Item onSelect={() => onSelect('Quote reply')}>
-              Quote reply
-              <ActionList.TrailingVisual>⌘Q</ActionList.TrailingVisual>
-            </ActionList.Item>
-            <ActionList.Item onSelect={() => onSelect('Edit comment')}>
-              Edit comment
-              <ActionList.TrailingVisual>⌘E</ActionList.TrailingVisual>
-            </ActionList.Item>
-            <ActionList.Divider />
-            <ActionList.Item variant="danger" onSelect={() => onSelect('Delete file')}>
-              Delete file
-              <ActionList.TrailingVisual>⌘D</ActionList.TrailingVisual>
-            </ActionList.Item>
-          </ActionList>
-        </ActionMenu.Overlay>
-      </ActionMenu>
-    </>
-  )
-}
-SimpleListStory.storyName = 'Simple Menu'
 
 export function ActionsStory(): JSX.Element {
   return (
